waves/aws/dynamic_node_pools_gke/requirements/requirements.txt
aws-lambda-typing==2.20.0
dacite
google-cloud-logging==3.11.3
templafirm>=0.0.13,<1
